ターン720度達成！記念

+++ src/info/keloud/leJOS/manager/DisplayUpdater.java
@@ -11,7 +11,7 @@ public class DisplayUpdater {
 
     public void updateValue(String operationMode, int accumulationleftMotor, int accumulationrightMotor, int accumulationcenterMotor, float colorIdValue, float ultrasonicValue, float gyroValue, int timer) {
         LCD.clear(0);
-        LCD.drawString(String.valueOf((float) ((int) (Battery.getVoltage() * 10 + 0.5) / 10.0)), 15, 0);
+        LCD.drawString(String.valueOf((float) ((Battery.getVoltage() * 10 + 0.5) / 10.0)), 15, 0);
         LCD.clear(1);
         LCD.drawString("L:" + accumulationleftMotor + " R:" + accumulationrightMotor + " C:" + accumulationcenterMotor, 1, 1);
         LCD.clear(2);
